+++ pkg/plan/plan.go
@@ -3,8 +3,6 @@ package plan
 
 import (
 	"github.com/spf13/afero"
-	"github.com/zeabur/zbpack/pkg/zeaburpack"
-
 	"github.com/zeabur/zbpack/pkg/types"
 )
 
@@ -22,7 +20,7 @@ type planner struct {
 // NewPlannerOptions is the options for NewPlanner.
 type NewPlannerOptions struct {
 	Source             afero.Fs
-	Config             zeaburpack.ImmutableProjectConfiguration
+	Config             ImmutableProjectConfiguration
 	SubmoduleName      string
 	CustomBuildCommand *string
 	CustomStartCommand *string
