+++ pkg/zeaburpack/plan.go
@@ -50,7 +50,7 @@ func Plan(opt PlanOptions) (types.PlanType, types.PlanMeta) {
 	planner := plan.NewPlanner(
 		&plan.NewPlannerOptions{
 			Source:             src,
-			Config:             NewProjectConfigurationFromFs(src),
+			Config:             plan.NewProjectConfigurationFromFs(src),
 			CustomBuildCommand: opt.CustomBuildCommand,
 			CustomStartCommand: opt.CustomStartCommand,
 			OutputDir:          opt.OutputDir,
